update proj4 source code and data to latest github master

+++ src/pj_gridlist.c
@@ -1,6 +1,4 @@
 /******************************************************************************
- * $Id: pj_gridlist.c 1990 2011-03-28 18:06:43Z warmerdam $
- *
  * Project:  PROJ.4
  * Purpose:  Code to manage the list of currently loaded (cached) PJ_GRIDINFOs
  *           See pj_gridinfo.c for details of loading individual grids.
@@ -178,7 +176,7 @@ PJ_GRIDINFO **pj_gridlist_from_nadgrids( projCtx ctx, const char *nadgrids,
 /* -------------------------------------------------------------------- */
     for( s = nadgrids; *s != '\0'; )
     {
-        int   end_char;
+        size_t end_char;
         int   required = 1;
         char  name[128];
